InsertSign: Fix chroma shift, Fix bitdepth conversion for the mask clip
In VapourSynth RGB clips have the frame property `_ChromaLocation` set to `1` (center). Converting the clip to YUV will set `_ChromaLocation` to `0` if not otherwise specified. However, `fmtc.matrix` will not set the property to `0`, so the internal resizer will treat the chroma as if it was center aligned. We now do RGB --> YUV conversion with the internal resizers, too.

Bitdepth conversion of the mask clip was flawed as full range wasn't assumed.

Make 709 the default matrix.